feat(react-x): consolidate prefer-set-state-callback into new unstable-rules-of-state rule, closes #1289
