+++ .pkgs/configs/package.json
@@ -22,8 +22,8 @@
   "dependencies": {
     "@stylistic/eslint-plugin": "^5.10.0",
     "eslint-plugin-de-morgan": "^2.1.1",
-    "eslint-plugin-function": "^0.9.0",
-    "eslint-plugin-function-rule": "^0.9.0",
+    "eslint-plugin-function": "^0.10.1",
+    "eslint-plugin-function-rule": "^0.10.1",
     "eslint-plugin-jsdoc": "^62.8.0",
     "eslint-plugin-perfectionist": "^5.6.0",
     "eslint-plugin-regexp": "^3.1.0",
@@ -33,7 +33,7 @@
   "devDependencies": {
     "eslint": "^10.0.3",
     "tsl": "^1.0.29",
-    "tsl-dx": "^0.9.0",
+    "tsl-dx": "^0.10.1",
     "typedoc": "0.28.17",
     "typedoc-plugin-markdown": "^4.10.0",
     "typedoc-plugin-mdn-links": "^5.1.1",
