+++ examples/preact/package.json
@@ -18,7 +18,7 @@
     "@preact/preset-vite": "^2.10.3",
     "@tsconfig/node24": "^24.0.4",
     "@tsconfig/strictest": "^2.0.8",
-    "@types/node": "^25.4.0",
+    "@types/node": "^25.5.0",
     "eslint": "^10.0.3",
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-react-refresh": "^0.5.2",
