+++ packages/plugins/eslint-plugin-react-x/src/configs/disable-experimental.ts
@@ -16,4 +16,5 @@ export const rules: Record<string, RuleConfig> = {
   "react-x/refs": "off",
   "react-x/set-state-in-render": "off",
   "react-x/unstable-rules-of-props": "off",
+  "react-x/unstable-rules-of-state": "off",
 };
