+++ packages/shared/package.json
@@ -45,7 +45,7 @@
     "@local/configs": "workspace:*",
     "@local/eff": "workspace:*",
     "@tsconfig/node24": "^24.0.4",
-    "@types/node": "^25.4.0",
+    "@types/node": "^25.5.0",
     "@types/picomatch": "^4.0.2",
     "tsdown": "^0.21.2"
   },
