+++ pnpm-workspace.yaml
@@ -21,7 +21,7 @@ overrides:
   "@typescript-eslint/types": "canary"
   "@typescript-eslint/utils": "canary"
   dompurify: ^3.3.3
-  esbuild: ^0.27.3
+  esbuild: ^0.27.4
   lodash: ^4.17.23
   minimatch: ^10.2.4
   next: ^16.1.6
@@ -32,7 +32,7 @@ overrides:
   tslib: ^2.8.1
   typescript-eslint: "canary"
   typescript: ^5.9.3
-  undici-types: ^7.23.0
-  undici: ^7.23.0
+  undici-types: ^7.24.0
+  undici: ^7.24.0
 
 trustPolicy: "no-downgrade"
